+++ CHANGELOG.md
@@ -8,10 +8,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added SSL connection capability to the Mongo classes.
 
 ### Added
+- mongo_class.Server.set_ssl_config:  Append SSL attributes to config.
 - mongo_class.Server.set_pass_config:  Set the passwd config attributes.
 
 ### Changed
-- mongo_class.Server.\_\_init\_\_:  Moved setting of config attribute to set_pass_config method.
+- mongo_class.Server.\_\_init\_\_:  Moved setting of config attribute to set_pass_config method and added SSL attributes and call.
 
 
 ## [4.1.0] - 2020-12-01